+++ packages/health/lib/src/health_factory.dart
@@ -14,9 +14,19 @@ class HealthFactory {
   static const MethodChannel _channel = MethodChannel('flutter_health');
   String? _deviceId;
   final _deviceInfo = DeviceInfoPlugin();
+  late bool _useHealthConnectIfAvailable;
 
   static PlatformType _platformType =
       Platform.isAndroid ? PlatformType.ANDROID : PlatformType.IOS;
+
+  // The plugin was created to use Health Connect (if true) or Google Fit (if false).
+  bool get useHealthConnectIfAvailable => _useHealthConnectIfAvailable;
+
+  HealthFactory({bool useHealthConnectIfAvailable = false}) {
+    _useHealthConnectIfAvailable = useHealthConnectIfAvailable;
+    if (_useHealthConnectIfAvailable)
+      _channel.invokeMethod('useHealthConnectIfAvailable');
+  }
 
   /// Check if a given data type is available on the platform
   bool isDataTypeAvailable(HealthDataType dataType) =>
@@ -47,7 +57,7 @@ class HealthFactory {
   ///   with a READ or READ_WRITE access.
   ///
   ///   On Android, this function returns true or false, depending on whether the specified access right has been granted.
-  static Future<bool?> hasPermissions(List<HealthDataType> types,
+  Future<bool?> hasPermissions(List<HealthDataType> types,
       {List<HealthDataAccess>? permissions}) async {
     if (permissions != null && permissions.length != types.length)
       throw ArgumentError(