+++ packages/health/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 5.0.0
+
+- Added initial support for the new Health Connect API, as Google Fit is being deprecated.
+  - Does not yet support `revokePermissions`, `getTotalStepsInInterval`.
+- Changed Intl package version dependancy to `^0.17.0` to work with flutter stable version.
+- Updated the example app to handle more buttons.
+
 ## 4.6.0
 
 - Added method for revoking permissions. On Android it uses `disableFit()` to remove access to Google Fit - `revokePermissions`. Documented lack of methods for iOS.